Add general modal for blocking the background
