+++ frontend/src/components/creator-modal/creator.tsx
@@ -3,6 +3,7 @@ import "./creator.scss";
 import { store } from "../../store";
 import CreatorEdit from "./creator-edit/creator-edit";
 import CreatorPreview from "./creator-preview/creator-preview";
+import Modal from "../modal/modal";
 
 //todo: create question props
 
@@ -34,48 +35,52 @@ const Creator = (): ReactElement => {
   };
 
   return (
-    <form className="create-question-module" onSubmit={saveQuestion}>
-      <div className="creator-header">
-        <button className="exit" onClick={closePreviewQuestion}>
-          ×
-        </button>
-        <span className="header-title">Create Question</span>
-        <div className="header-tabs">
-          <div
-            className={`tab-buttons edit-tab ${isPreview ? "" : "selected"}`}
-            onClick={() => setIsPreview(false)}
-          >
-            Edit
-          </div>
-          <div
-            className={`tab-buttons preview-tab ${isPreview ? "selected" : ""}`}
-            onClick={() => setIsPreview(true)}
-          >
-            Preview
+    <Modal>
+      <form className="create-question-module" onSubmit={saveQuestion}>
+        <div className="creator-header">
+          <button className="exit" onClick={closePreviewQuestion}>
+            ×
+          </button>
+          <span className="header-title">Create Question</span>
+          <div className="header-tabs">
+            <div
+              className={`tab-buttons edit-tab ${isPreview ? "" : "selected"}`}
+              onClick={() => setIsPreview(false)}
+            >
+              Edit
+            </div>
+            <div
+              className={`tab-buttons preview-tab ${
+                isPreview ? "selected" : ""
+              }`}
+              onClick={() => setIsPreview(true)}
+            >
+              Preview
+            </div>
           </div>
         </div>
-      </div>
-      {isPreview ? (
-        <CreatorPreview newQuestion={newQuestion} />
-      ) : (
-        <CreatorEdit
-          newQuestion={newQuestion}
-          setNewQuestion={setNewQuestion}
-        />
-      )}
-      <div className="buttons">
-        <button
-          type="reset"
-          className="cancel-button"
-          onClick={closePreviewQuestion}
-        >
-          Cancel
-        </button>
-        <button type="submit" className="save-button">
-          Save
-        </button>
-      </div>
-    </form>
+        {isPreview ? (
+          <CreatorPreview newQuestion={newQuestion} />
+        ) : (
+          <CreatorEdit
+            newQuestion={newQuestion}
+            setNewQuestion={setNewQuestion}
+          />
+        )}
+        <div className="buttons">
+          <button
+            type="reset"
+            className="cancel-button"
+            onClick={closePreviewQuestion}
+          >
+            Cancel
+          </button>
+          <button type="submit" className="save-button">
+            Save
+          </button>
+        </div>
+      </form>
+    </Modal>
   );
 };
