+++ frontend/src/components/export-modal/export-modal.tsx
@@ -6,6 +6,7 @@ import data from "./mock-data.json";
 
 import "./export-modal.scss";
 import SessionDropdown from "./session-dropdown/session-dropdown";
+import Modal from "../modal/modal";
 
 const ExportModal = (): ReactElement => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -23,62 +24,68 @@ const ExportModal = (): ReactElement => {
   };
 
   return (
-    <form className="export-modal">
-      <div className="export-header">
-        <h3>Export to Webcourses</h3>
+    <Modal>
+      <form className="export-modal">
+        <div className="export-header">
+          <h3>Export to Webcourses</h3>
 
-        <button className="exit-button" onClick={cancel}>
-          X
-        </button>
-      </div>
-
-      <div className="session-list">
-        <div className="labels">
-          <span>Session</span>
-
-          <span className="points">Points</span>
-
-          <span className="date">Date</span>
+          <button className="exit-button" onClick={cancel}>
+            X
+          </button>
         </div>
 
-        {sessions.map(
-          (session: Session, index: number): ReactElement => (
-            <SessionDropdown
-              key={index}
-              name={session.name}
-              date={session.date}
-              questionTitles={session.questions.title}
-              points={session.total}
-              index={index}
+        <div className="session-list">
+          <div className="labels">
+            <span>Session</span>
+
+            <span className="points">Points</span>
+
+            <span className="date">Date</span>
+          </div>
+
+          {sessions.map(
+            (session: Session, index: number): ReactElement => (
+              <SessionDropdown
+                key={index}
+                name={session.name}
+                date={session.date}
+                questionTitles={session.questions.title}
+                points={session.total}
+                index={index}
+              />
+            )
+          )}
+        </div>
+
+        <div className="metadata">
+          <div className="assignment-input">
+            <label htmlFor="export-assignment-name">Assignment Name:</label>
+
+            <input id="export-assignment-name" placeholder="Assignment 1" />
+          </div>
+
+          <div>
+            <label htmlFor="export-assignment-points">Assignment Points:</label>
+
+            <input
+              id="export-assignment-points"
+              type="number"
+              placeholder="10"
             />
-          )
-        )}
-      </div>
-
-      <div className="metadata">
-        <div className="assignment-input">
-          <label htmlFor="export-assignment-name">Assignment Name:</label>
-
-          <input id="export-assignment-name" placeholder="Assignment 1" />
+          </div>
         </div>
 
-        <div>
-          <label htmlFor="export-assignment-points">Assignment Points:</label>
+        <div className="export-footer">
+          <button className="cancel" onClick={cancel}>
+            Cancel
+          </button>
 
-          <input id="export-assignment-points" type="number" placeholder="10" />
+          <button className="export" onClick={exportGrades}>
+            Export
+          </button>
         </div>
-      </div>
-
-      <div className="export-footer">
-        <button className="cancel" onClick={cancel}>
-          Cancel
-        </button>
-
-        <button className="export" onClick={exportGrades}>
-          Export
-        </button>
-      </div>
-    </form>
+      </form>
+    </Modal>
   );
 };
